requirements.txt
pillow
python-telegram-bot
RPi.GPIO
python-dotenv
requests